chain, core, graph, mock: Simplify `EthereumAdapter`
Specifically, pull out the `Arc<Self>` methods from the trait so it
becomes easier to mock the trait with mockall (see the next commit).
