+++ chain/ethereum/src/block_stream.rs
@@ -4,6 +4,7 @@ use std::mem;
 use std::sync::Mutex;
 use std::time::{Duration, Instant};
 
+use graph::components::ethereum::{blocks_with_triggers, triggers_in_block};
 use graph::data::subgraph::schema::{
     SubgraphDeploymentEntity, SubgraphEntity, SubgraphVersionEntity,
 };
@@ -423,23 +424,21 @@ where
                                 "range_size" => range_size
                             );
                             Box::new(
-                                ctx.eth_adapter
-                                    .blocks_with_triggers(
-                                        ctx.logger.clone(),
-                                        ctx.chain_store.clone(),
-                                        ctx.metrics.ethrpc_metrics.clone(),
-                                        from,
-                                        to,
-                                        log_filter.clone(),
-                                        call_filter.clone(),
-                                        block_filter.clone(),
-                                    )
-                                    .map(move |blocks| {
-                                        section.end();
-                                        ReconciliationStep::ProcessDescendantBlocks(
-                                            blocks, range_size,
-                                        )
-                                    }),
+                                blocks_with_triggers(
+                                    ctx.eth_adapter,
+                                    ctx.logger.clone(),
+                                    ctx.chain_store.clone(),
+                                    ctx.metrics.ethrpc_metrics.clone(),
+                                    from,
+                                    to,
+                                    log_filter.clone(),
+                                    call_filter.clone(),
+                                    block_filter.clone(),
+                                )
+                                .map(move |blocks| {
+                                    section.end();
+                                    ReconciliationStep::ProcessDescendantBlocks(blocks, range_size)
+                                }),
                             )
                         },
                     ),
@@ -520,7 +519,8 @@ where
                         Box::new(
                             block_with_calls
                                 .and_then(move |block| {
-                                    eth_adapter.triggers_in_block(
+                                    triggers_in_block(
+                                        eth_adapter,
                                         logger,
                                         ctx.chain_store.clone(),
                                         ctx.metrics.ethrpc_metrics.clone(),
